pyproject.toml
[tool.poetry]
name = "lore_log"
version = "0.1.0"
description = "Backend API Service for Lore Log"
authors = ["Zach Bullough <zbullough@qmulosoft.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
falcon = "^2.0.0"
falcon-auth = "^1.1.0"
bcrypt = "^3.1.7"
pyjwt = "^1.7.1"
nimbus-migrate = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
